Unify the memory management api in client
For now, client.prepare_device is the only
api to make room for new chunks.
This commit also remove the redundant code after
this change.

+++ examples/pretrain_demo.py
@@ -242,7 +242,7 @@ if __name__ == "__main__":
         print(f"patrickstar: {ps_res_list}")
 
         def diff(array):
-            dtype = np.float16 if use_fp16 else np.float
+            dtype = np.float16 if use_fp16 else float
             return list(
                 np.array(ps_res_list, dtype=dtype) - np.array(array, dtype=dtype)
             )
